agente-reclutamiento-v0/requirements.txt
# ============================================
# Dependencias del sistema de extracción de CVs
# ============================================

# Extracción de PDFs
pdfplumber>=0.10.0

# Extracción de Word (.docx)
python-docx>=1.1.0

# Automatización COM para .doc (solo Windows)
pywin32>=306

# Google Gemini SDK (nuevo)
google-genai>=0.3.0

# Procesamiento de imágenes
Pillow>=10.0.0

# Exportación a Excel
pandas>=2.0.0
openpyxl>=3.1.0

# Configuración
python-dotenv>=1.0.0
